requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
Pillow>=10.3.0
numpy>=1.26.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
torch>=2.2.0
transformers>=4.40.0
huggingface_hub>=0.23.0
